% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_acar.R
\name{compute_acar}
\alias{compute_acar}
\title{Compute ACAR}
\usage{
compute_acar(factor_data, forward_returns, k, ntiles = 5)
}
\arguments{
\item{factor_data}{xts object}

\item{forward_returns}{xts object}

\item{k}{integer forward returns window in months}

\item{ntiles}{integer}
}
\description{
See Bondt & Thaler 1985
}
