% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sort_portfolios.R
\name{sort_portfolios}
\alias{sort_portfolios}
\title{Characteristic-Sorted Portfolios}
\usage{
sort_portfolios(factor_data, forward_returns, k, ntiles = 5)
}
\arguments{
\item{factor_data}{xts object}

\item{forward_returns}{xts object}

\item{k}{integer forward returns window in months}

\item{ntiles}{integer}
}
\value{
list with monthly rankings and portfolios returns
}
\description{
Construct characteristic-sorted portfolios
}
\note{
Formation months are extracted from factor_data.

Formation date are the beginning of each month.

Thus, forward_returns must include at least (k-1) months of leading data.
}
